Changed data type for n from int to long 7-print_last_digit.c

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -5,7 +5,7 @@
   *
   * Return: Last digit of the given n
   */
-int print_last_digit(long n)
+long print_last_digit(long n)
 {
 	if (n < 0)
 	{
